remove unnecessary const values and create ring/kv/mock.go
