+++ pkg/ring/kv/client.go
@@ -6,14 +6,12 @@ import (
 	"fmt"
 	"sync"
 
-	"github.com/go-kit/kit/log/level"
 	"github.com/prometheus/client_golang/prometheus"
 
 	"github.com/cortexproject/cortex/pkg/ring/kv/codec"
 	"github.com/cortexproject/cortex/pkg/ring/kv/consul"
 	"github.com/cortexproject/cortex/pkg/ring/kv/etcd"
 	"github.com/cortexproject/cortex/pkg/ring/kv/memberlist"
-	"github.com/cortexproject/cortex/pkg/util"
 )
 
 const (
@@ -207,35 +205,4 @@ func buildMultiClient(cfg StoreConfig, codec codec.Codec, reg prometheus.Registe
 	}
 
 	return NewMultiClient(cfg.Multi, clients), nil
-}
-
-// The mockClient does not anything.
-// This is used for testing only.
-type mockClient struct{}
-
-func buildMockClient() (Client, error) {
-	level.Warn(util.Logger).Log("msg", "created mockClient for testing only")
-	return mockClient{}, nil
-}
-
-func (m mockClient) List(ctx context.Context, prefix string) ([]string, error) {
-	return []string{}, nil
-}
-
-func (m mockClient) Get(ctx context.Context, key string) (interface{}, error) {
-	return "", nil
-}
-
-func (m mockClient) Delete(ctx context.Context, key string) error {
-	return nil
-}
-
-func (m mockClient) CAS(ctx context.Context, key string, f func(in interface{}) (out interface{}, retry bool, err error)) error {
-	return nil
-}
-
-func (m mockClient) WatchKey(ctx context.Context, key string, f func(interface{}) bool) {
-}
-
-func (m mockClient) WatchPrefix(ctx context.Context, prefix string, f func(string, interface{}) bool) {
 }
