+++ pkg/ring/kv/metrics.go
@@ -31,12 +31,6 @@ func errorCode(err error) string {
 	return "500"
 }
 
-const (
-	MetricNamespace = "cortex"
-	MetricName      = "kv_request_duration_seconds"
-	MetricHelp      = "Time spent on kv store requests."
-)
-
 type metrics struct {
 	c               Client
 	requestDuration *instrument.HistogramCollector
@@ -47,9 +41,9 @@ func newMetricsClient(backend string, c Client, reg prometheus.Registerer) Clien
 		c: c,
 		requestDuration: instrument.NewHistogramCollector(
 			promauto.With(reg).NewHistogramVec(prometheus.HistogramOpts{
-				Namespace: MetricNamespace,
-				Name:      MetricName,
-				Help:      MetricHelp,
+				Namespace: "cortex",
+				Name:      "kv_request_duration_seconds",
+				Help:      "Time spent on kv store requests.",
 				Buckets:   prometheus.DefBuckets,
 				ConstLabels: prometheus.Labels{
 					"type": backend,
